# openapi_complete_v5.3.yaml
openapi: 3.1.0
info:
  title: DentalKiosk Shopware E-Commerce API
  description: 'Complete API for the DentalKiosk chatbot middleware used with Shopware
    6 and Dify.


    Features:

    - Product search & details

    - Cart action queue (add/remove/update/clear)

    - Orders lookup and reorder (by orderNumber)

    - Multimodal pipeline (image -> extract -> search -> top3 -> confirm)

    - Analytics ingestion + summary

    '
  version: 5.3.0
  contact:
    name: DentalKiosk Support
    url: https://dentalkiosk.de
servers:
- url: https://cbshop.commerce-crew.com
  description: Production Middleware Server
- url: http://localhost:3330
  description: Local Development
...
paths:
...
  /api/analytics:
    post:
      operationId: ingestAnalytics
      summary: Ingest analytics events batch
      tags:
      - Analytics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - events
              properties:
                events:
                  type: array
                  items:
                    type: object
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  received:
                    type: integer
    get:
      operationId: getAnalytics
      summary: Get aggregated analytics and recent events
      tags:
      - Analytics
      responses:
        '200':
          description: Aggregated analytics
          content:
            application/json:
              schema:
                type: object
  /api/analytics/questions:
    get:
      operationId: getAnalyticsQuestions
      summary: Get recent questions and top repeated questions
      tags:
      - Analytics
      parameters:
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
      responses:
        '200':
          description: Questions
          content:
            application/json:
              schema:
                type: object
...